--- a/etbrasil/geesebal/__init__.py
+++ b/etbrasil/geesebal/__init__.py
@@ -1,5 +1,5 @@
 from .image import Image
 from .collection import Collection
-from .timeseries import TimeSeries
+from .timeseries import TimeSeries, TimeSeriesAsync
 
-__version__ = "0.1.1"
+__version__ = "0.1.2"